supersonic timing and including some more files to make it easier to clone.
