+++ KickoffTimerPlugin.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29806.167
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "KickoffTimerPlugin", "KickoffTimerPlugin\KickoffTimerPlugin.vcxproj", "{2F242864-84C1-4181-9E1E-02BB3AF89818}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2F242864-84C1-4181-9E1E-02BB3AF89818}.Debug|x64.ActiveCfg = Debug|x64
+		{2F242864-84C1-4181-9E1E-02BB3AF89818}.Debug|x64.Build.0 = Debug|x64
+		{2F242864-84C1-4181-9E1E-02BB3AF89818}.Release|x64.ActiveCfg = Release|x64
+		{2F242864-84C1-4181-9E1E-02BB3AF89818}.Release|x64.Build.0 = Release|x64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {3E7FA28A-6CB9-4CA9-90BC-BCB737554BEB}
+	EndGlobalSection
+EndGlobal
