+++ KickoffTimerPlugin/pch.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#define WIN32_LEAN_AND_MEAN
+#define _CRT_SECURE_NO_WARNINGS
+#include "bakkesmod/plugin/bakkesmodplugin.h"
+
+#include <string>
+#include <vector>
+#include <functional>
+#include <memory>
